refactor: change club tag endpoints
