+++ src/main/java/kr/hanjari/backend/domain/tag/presentation/controller/TagController.java
@@ -10,9 +10,7 @@ import kr.hanjari.backend.domain.tag.presentation.dto.response.GetAllTagsRespons
 import kr.hanjari.backend.domain.tag.presentation.dto.response.TagIdResponse;
 import kr.hanjari.backend.global.payload.ApiResponse;
 import lombok.RequiredArgsConstructor;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -45,29 +43,5 @@ public class TagController {
     public ApiResponse<TagIdResponse> createTag(@RequestBody @Valid CreateTagRequest request) {
         Long tagId = tagCommandService.createTag(request.name(), request.priority());
         return ApiResponse.onSuccess(TagIdResponse.of(tagId));
-    }
-
-    @Operation(summary = "[태그] 동아리 태그 추가", description = """
-            ## 동아리에 태그를 추가합니다.
-            ### Path Variables
-            - **clubId**: 태그를 추가할 동아리 ID
-            - **tagId**: 추가할 태그 ID
-            """)
-    @PostMapping("/clubs/{clubId}/{tagId}")
-    public ApiResponse<Void> addTagToClub(@PathVariable Long clubId, @PathVariable Long tagId) {
-        tagCommandService.addTagToClub(clubId, tagId);
-        return ApiResponse.onSuccess();
-    }
-
-    @Operation(summary = "[태그] 동아리 태그 삭제", description = """
-            ## 동아리에서 태그를 삭제합니다.
-            ### Path Variables
-            - **clubId**: 태그를 삭제할 동아리 ID
-            - **tagId**: 삭제할 태그 ID
-            """)
-    @DeleteMapping("/clubs/{clubId}/{tagId}")
-    public ApiResponse<Void> removeTagFromClub(@PathVariable Long clubId, @PathVariable Long tagId) {
-        tagCommandService.removeTagFromClub(clubId, tagId);
-        return ApiResponse.onSuccess();
     }
 }
